+++ src/main/java/net/rubyeye/xmemcached/command/binary/BinaryQuitCommand.java
@@ -70,7 +70,7 @@ public class BinaryQuitCommand extends BaseBinaryCommand {
 	}
 
 	@Override
-	protected int getKeyLength() {
+	protected short getKeyLength() {
 		return 0;
 	}
