+++ src/main/java/net/rubyeye/xmemcached/command/binary/BinaryVerbosityCommand.java
@@ -39,7 +39,7 @@ public class BinaryVerbosityCommand extends BaseBinaryCommand {
 	}
 
 	@Override
-	protected int getKeyLength() {
+	protected short getKeyLength() {
 		return 0;
 	}
